feat: combine basic and functional test suites into one
Merge the functional ASG scale-in tests into the basic suite to run
everything in a single test cluster, reducing CI cost and time.
