--- tests/e2e/suites/functional/config.yaml
@@ -1,9 +1,0 @@
-appName: "aws-node-termination-handler-bundle"
-repoName: "aws-node-termination-handler-app"
-appCatalog: "default"
-
-providers:
-  - capa
-
-helmReleaseSourceRef:
-  name: default-test-catalog
