--- tests/e2e/suites/functional/values.yaml
@@ -1,1 +1,0 @@
-# Default bundle values are sufficient for functional test
